move grpc create table to sql crates
