+++ server/src/grpc/storage_service/mod.rs
@@ -3,7 +3,7 @@
 // Storage rpc service implement.
 
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::HashMap,
     stringify,
     sync::Arc,
     time::{Duration, Instant},
@@ -11,15 +11,8 @@ use std::{
 
 use async_trait::async_trait;
 use ceresdbproto::storage::{
-    storage_service_server::StorageService, value::Value, PrometheusQueryRequest,
-    PrometheusQueryResponse, RouteRequest, RouteResponse, SqlQueryRequest, SqlQueryResponse,
-    WriteRequest, WriteResponse, WriteTableRequest,
-};
-use cluster::config::SchemaConfig;
-use common_types::{
-    column_schema::{self, ColumnSchema},
-    datum::DatumKind,
-    schema::{Builder as SchemaBuilder, Schema, TSID_COLUMN},
+    storage_service_server::StorageService, PrometheusQueryRequest, PrometheusQueryResponse,
+    RouteRequest, RouteResponse, SqlQueryRequest, SqlQueryResponse, WriteRequest, WriteResponse,
 };
 use common_util::{error::BoxError, time::InstantExt};
 use futures::stream::{self, BoxStream, StreamExt};
@@ -29,8 +22,7 @@ use log::{error, warn};
 use paste::paste;
 use query_engine::executor::Executor as QueryExecutor;
 use router::{Router, RouterRef};
-use snafu::{ensure, OptionExt, ResultExt};
-use sql::plan::CreateTablePlan;
+use snafu::ResultExt;
 use table_engine::engine::EngineRuntimes;
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
@@ -447,244 +439,6 @@ impl<Q: QueryExecutor + 'static> StorageService for StorageServiceImpl<Q> {
                 Ok(tonic::Response::new(Box::pin(stream)))
             }
         }
-    }
-}
-
-/// Create CreateTablePlan from a write metric.
-///
-/// The caller must ENSURE that the HandlerContext's schema_config is not None.
-pub fn write_table_request_to_create_table_plan(
-    schema_config: &SchemaConfig,
-    write_table: &WriteTableRequest,
-) -> Result<CreateTablePlan> {
-    Ok(CreateTablePlan {
-        engine: schema_config.default_engine_type.clone(),
-        if_not_exists: true,
-        table: write_table.table.clone(),
-        table_schema: build_schema_from_write_table_request(schema_config, write_table)?,
-        options: HashMap::default(),
-        partition_info: None,
-    })
-}
-
-fn build_column_schema(
-    column_name: &str,
-    data_type: DatumKind,
-    is_tag: bool,
-) -> Result<ColumnSchema> {
-    let builder = column_schema::Builder::new(column_name.to_string(), data_type)
-        .is_nullable(true)
-        .is_tag(is_tag);
-
-    builder.build().box_err().context(ErrWithCause {
-        code: StatusCode::BAD_REQUEST,
-        msg: "invalid column schema",
-    })
-}
-
-fn build_schema_from_write_table_request(
-    schema_config: &SchemaConfig,
-    write_table_req: &WriteTableRequest,
-) -> Result<Schema> {
-    let WriteTableRequest {
-        table,
-        field_names,
-        tag_names,
-        entries: write_entries,
-    } = write_table_req;
-
-    let mut schema_builder =
-        SchemaBuilder::with_capacity(field_names.len()).auto_increment_column_id(true);
-
-    ensure!(
-        !write_entries.is_empty(),
-        ErrNoCause {
-            code: StatusCode::BAD_REQUEST,
-            msg: format!("empty write entires to write table:{table}"),
-        }
-    );
-
-    let mut name_column_map: BTreeMap<_, ColumnSchema> = BTreeMap::new();
-    for write_entry in write_entries {
-        // parse tags
-        for tag in &write_entry.tags {
-            let name_index = tag.name_index as usize;
-            ensure!(
-                name_index < tag_names.len(),
-                ErrNoCause {
-                    code: StatusCode::BAD_REQUEST,
-                    msg: format!(
-                        "tag index {name_index} is not found in tag_names:{tag_names:?}, table:{table}",
-                    ),
-                }
-            );
-
-            let tag_name = &tag_names[name_index];
-
-            let tag_value = tag
-                .value
-                .as_ref()
-                .with_context(|| ErrNoCause {
-                    code: StatusCode::BAD_REQUEST,
-                    msg: format!("Tag({tag_name}) value is needed, table_name:{table} "),
-                })?
-                .value
-                .as_ref()
-                .with_context(|| ErrNoCause {
-                    code: StatusCode::BAD_REQUEST,
-                    msg: format!("Tag({tag_name}) value type is not supported, table_name:{table}"),
-                })?;
-
-            let data_type = try_get_data_type_from_value(tag_value)?;
-
-            if let Some(column_schema) = name_column_map.get(tag_name) {
-                ensure_data_type_compatible(table, tag_name, true, data_type, column_schema)?;
-            }
-            let column_schema = build_column_schema(tag_name, data_type, true)?;
-            name_column_map.insert(tag_name, column_schema);
-        }
-
-        // parse fields
-        for field_group in &write_entry.field_groups {
-            for field in &field_group.fields {
-                if (field.name_index as usize) < field_names.len() {
-                    let field_name = &field_names[field.name_index as usize];
-                    let field_value = field
-                        .value
-                        .as_ref()
-                        .with_context(|| ErrNoCause {
-                            code: StatusCode::BAD_REQUEST,
-                            msg: format!("Field({field_name}) value is needed, table:{table}"),
-                        })?
-                        .value
-                        .as_ref()
-                        .with_context(|| ErrNoCause {
-                            code: StatusCode::BAD_REQUEST,
-                            msg: format!(
-                                "Field({field_name}) value type is not supported, table:{table}"
-                            ),
-                        })?;
-
-                    let data_type = try_get_data_type_from_value(field_value)?;
-
-                    if let Some(column_schema) = name_column_map.get(field_name) {
-                        ensure_data_type_compatible(
-                            table,
-                            field_name,
-                            false,
-                            data_type,
-                            column_schema,
-                        )?;
-                    }
-
-                    let column_schema = build_column_schema(field_name, data_type, false)?;
-                    name_column_map.insert(field_name, column_schema);
-                }
-            }
-        }
-    }
-
-    // Timestamp column will be the last column
-    let timestamp_column_schema = column_schema::Builder::new(
-        schema_config.default_timestamp_column_name.clone(),
-        DatumKind::Timestamp,
-    )
-    .is_nullable(false)
-    .build()
-    .box_err()
-    .context(ErrWithCause {
-        code: StatusCode::BAD_REQUEST,
-        msg: "invalid timestamp column schema to build",
-    })?;
-
-    // Use (tsid, timestamp) as primary key.
-    let tsid_column_schema =
-        column_schema::Builder::new(TSID_COLUMN.to_string(), DatumKind::UInt64)
-            .is_nullable(false)
-            .build()
-            .box_err()
-            .context(ErrWithCause {
-                code: StatusCode::BAD_REQUEST,
-                msg: "invalid tsid column schema to build",
-            })?;
-
-    schema_builder = schema_builder
-        .add_key_column(tsid_column_schema)
-        .box_err()
-        .context(ErrWithCause {
-            code: StatusCode::BAD_REQUEST,
-            msg: "invalid timestamp column to add",
-        })?
-        .add_key_column(timestamp_column_schema)
-        .box_err()
-        .context(ErrWithCause {
-            code: StatusCode::BAD_REQUEST,
-            msg: "invalid tsid column to add",
-        })?;
-
-    for col in name_column_map.into_values() {
-        schema_builder = schema_builder
-            .add_normal_column(col)
-            .box_err()
-            .context(ErrWithCause {
-                code: StatusCode::BAD_REQUEST,
-                msg: "invalid normal column to add",
-            })?;
-    }
-
-    schema_builder.build().box_err().context(ErrWithCause {
-        code: StatusCode::BAD_REQUEST,
-        msg: "invalid schema to build",
-    })
-}
-
-fn ensure_data_type_compatible(
-    table_name: &str,
-    column_name: &str,
-    is_tag: bool,
-    data_type: DatumKind,
-    column_schema: &ColumnSchema,
-) -> Result<()> {
-    ensure!(
-        column_schema.is_tag == is_tag,
-        ErrNoCause {
-            code: StatusCode::BAD_REQUEST,
-            msg: format!(
-                "Duplicated column: {column_name} in fields and tags for table: {table_name}",
-            ),
-        }
-    );
-
-    ensure!(
-        column_schema.data_type == data_type,
-        ErrNoCause {
-            code: StatusCode::BAD_REQUEST,
-            msg: format!(
-                "Column: {} in table: {} data type is not same, expected: {}, actual: {}",
-                column_name, table_name, column_schema.data_type, data_type,
-            ),
-        }
-    );
-
-    Ok(())
-}
-
-fn try_get_data_type_from_value(value: &Value) -> Result<DatumKind> {
-    match value {
-        Value::Float64Value(_) => Ok(DatumKind::Double),
-        Value::StringValue(_) => Ok(DatumKind::String),
-        Value::Int64Value(_) => Ok(DatumKind::Int64),
-        Value::Float32Value(_) => Ok(DatumKind::Float),
-        Value::Int32Value(_) => Ok(DatumKind::Int32),
-        Value::Int16Value(_) => Ok(DatumKind::Int16),
-        Value::Int8Value(_) => Ok(DatumKind::Int8),
-        Value::BoolValue(_) => Ok(DatumKind::Boolean),
-        Value::Uint64Value(_) => Ok(DatumKind::UInt64),
-        Value::Uint32Value(_) => Ok(DatumKind::UInt32),
-        Value::Uint16Value(_) => Ok(DatumKind::UInt16),
-        Value::Uint8Value(_) => Ok(DatumKind::UInt8),
-        Value::TimestampValue(_) => Ok(DatumKind::Timestamp),
-        Value::VarbinaryValue(_) => Ok(DatumKind::Varbinary),
     }
 }
