+++ sql/Cargo.toml
@@ -19,6 +19,7 @@ arrow = { workspace = true }
 async-trait = { workspace = true }
 catalog = { workspace = true }
 ceresdbproto = { workspace = true }
+cluster = { workspace = true }
 common_types = { workspace = true }
 common_util = { workspace = true }
 datafusion = { workspace = true }
