+++ sql/src/planner.rs
@@ -16,12 +16,14 @@ use arrow::{
     error::ArrowError,
 };
 use catalog::consts::{DEFAULT_CATALOG, DEFAULT_SCHEMA};
+use ceresdbproto::storage::{value::Value as PbValue, WriteTableRequest};
+use cluster::config::SchemaConfig;
 use common_types::{
     column_schema::{self, ColumnSchema},
     datum::{Datum, DatumKind},
     request_id::RequestId,
     row::{RowGroup, RowGroupBuilder},
-    schema::{self, Schema, TSID_COLUMN},
+    schema::{self, Builder as SchemaBuilder, Schema, TSID_COLUMN},
 };
 use common_util::error::GenericError;
 use datafusion::{
@@ -247,6 +249,9 @@ pub enum Error {
 
     #[snafu(display("Unsupported partition method, msg:{}", msg,))]
     UnsupportedPartition { msg: String },
+
+    #[snafu(display("Failed to build plan, msg:{}", msg))]
+    InvalidWriteEntry { msg: String },
 }
 
 define_result!(Error);
@@ -265,7 +270,7 @@ pub struct Planner<'a, P: MetaProvider> {
     read_parallelism: usize,
 }
 
-impl<'a, P: MetaProvider> Planner<'a, P> {
+impl<'a, P: MetaProvider + 'a> Planner<'a, P> {
     /// Create a new logical planner
     pub fn new(provider: &'a P, request_id: RequestId, read_parallelism: usize) -> Self {
         Self {
@@ -317,6 +322,223 @@ impl<'a, P: MetaProvider> Planner<'a, P> {
 
         expr.to_plan(planner.meta_provider, self.read_parallelism)
             .context(BuildPromPlanError)
+    }
+
+    pub fn write_req_to_plan(
+        &self,
+        schema_config: &SchemaConfig,
+        write_table: &WriteTableRequest,
+    ) -> Result<Plan> {
+        Ok(Plan::Create(CreateTablePlan {
+            engine: schema_config.default_engine_type.clone(),
+            if_not_exists: true,
+            table: write_table.table.clone(),
+            table_schema: Self::build_schema_from_write_table_request(schema_config, write_table)?,
+            options: HashMap::default(),
+            partition_info: None,
+        }))
+    }
+
+    fn build_column_schema(
+        column_name: &str,
+        data_type: DatumKind,
+        is_tag: bool,
+    ) -> Result<ColumnSchema> {
+        let builder = column_schema::Builder::new(column_name.to_string(), data_type)
+            .is_nullable(true)
+            .is_tag(is_tag);
+
+        builder.build().with_context(|| InvalidColumnSchema {
+            column_name: column_name.to_string(),
+        })
+    }
+
+    fn build_schema_from_write_table_request(
+        schema_config: &SchemaConfig,
+        write_table_req: &WriteTableRequest,
+    ) -> Result<Schema> {
+        let WriteTableRequest {
+            table,
+            field_names,
+            tag_names,
+            entries: write_entries,
+        } = write_table_req;
+
+        let mut schema_builder =
+            SchemaBuilder::with_capacity(field_names.len()).auto_increment_column_id(true);
+
+        ensure!(
+            !write_entries.is_empty(),
+            InvalidWriteEntry {
+                msg: "empty write entries".to_string()
+            }
+        );
+
+        let mut name_column_map: BTreeMap<_, ColumnSchema> = BTreeMap::new();
+        for write_entry in write_entries {
+            // parse tags
+            for tag in &write_entry.tags {
+                let name_index = tag.name_index as usize;
+                ensure!(
+                name_index < tag_names.len(),
+                InvalidWriteEntry{
+                    msg: format!(
+                        "tag index {name_index} is not found in tag_names:{tag_names:?}, table:{table}",
+                    ),
+                }
+            );
+
+                let tag_name = &tag_names[name_index];
+
+                let tag_value = tag
+                    .value
+                    .as_ref()
+                    .with_context(|| InvalidWriteEntry {
+                        msg: format!("Tag({tag_name}) value is needed, table_name:{table} "),
+                    })?
+                    .value
+                    .as_ref()
+                    .with_context(|| InvalidWriteEntry {
+                        msg: format!(
+                            "Tag({tag_name}) value type is not supported, table_name:{table}"
+                        ),
+                    })?;
+
+                let data_type = Self::try_get_data_type_from_value(tag_value)?;
+
+                if let Some(column_schema) = name_column_map.get(tag_name) {
+                    Self::ensure_data_type_compatible(
+                        table,
+                        tag_name,
+                        true,
+                        data_type,
+                        column_schema,
+                    )?;
+                }
+                let column_schema = Self::build_column_schema(tag_name, data_type, true)?;
+                name_column_map.insert(tag_name, column_schema);
+            }
+
+            // parse fields
+            for field_group in &write_entry.field_groups {
+                for field in &field_group.fields {
+                    if (field.name_index as usize) < field_names.len() {
+                        let field_name = &field_names[field.name_index as usize];
+                        let field_value = field
+                            .value
+                            .as_ref()
+                            .with_context(|| InvalidWriteEntry {
+                                msg: format!("Field({field_name}) value is needed, table:{table}"),
+                            })?
+                            .value
+                            .as_ref()
+                            .with_context(|| InvalidWriteEntry {
+                                msg: format!(
+                                "Field({field_name}) value type is not supported, table:{table}"
+                            ),
+                            })?;
+
+                        let data_type = Self::try_get_data_type_from_value(field_value)?;
+
+                        if let Some(column_schema) = name_column_map.get(field_name) {
+                            Self::ensure_data_type_compatible(
+                                table,
+                                field_name,
+                                false,
+                                data_type,
+                                column_schema,
+                            )?;
+                        }
+
+                        let column_schema =
+                            Self::build_column_schema(field_name, data_type, false)?;
+                        name_column_map.insert(field_name, column_schema);
+                    }
+                }
+            }
+        }
+
+        // Timestamp column will be the last column
+        let timestamp_column_schema = column_schema::Builder::new(
+            schema_config.default_timestamp_column_name.clone(),
+            DatumKind::Timestamp,
+        )
+        .is_nullable(false)
+        .build()
+        .with_context(|| InvalidColumnSchema {
+            column_name: schema_config.default_timestamp_column_name.clone(),
+        })?;
+
+        // Use (tsid, timestamp) as primary key.
+        let tsid_column_schema =
+            column_schema::Builder::new(TSID_COLUMN.to_string(), DatumKind::UInt64)
+                .is_nullable(false)
+                .build()
+                .with_context(|| InvalidColumnSchema {
+                    column_name: TSID_COLUMN.to_string(),
+                })?;
+
+        schema_builder = schema_builder
+            .add_key_column(tsid_column_schema)
+            .context(BuildTableSchema {})?
+            .add_key_column(timestamp_column_schema)
+            .context(BuildTableSchema {})?;
+
+        for col in name_column_map.into_values() {
+            schema_builder = schema_builder
+                .add_normal_column(col)
+                .context(BuildTableSchema {})?;
+        }
+
+        schema_builder.build().context(BuildTableSchema {})
+    }
+
+    fn ensure_data_type_compatible(
+        table_name: &str,
+        column_name: &str,
+        is_tag: bool,
+        data_type: DatumKind,
+        column_schema: &ColumnSchema,
+    ) -> Result<()> {
+        ensure!(
+            column_schema.is_tag == is_tag,
+            InvalidWriteEntry {
+                msg: format!(
+                    "Duplicated column: {column_name} in fields and tags for table: {table_name}",
+                ),
+            }
+        );
+
+        ensure!(
+            column_schema.data_type == data_type,
+            InvalidWriteEntry {
+                msg: format!(
+                    "Column: {} in table: {} data type is not same, expected: {}, actual: {}",
+                    column_name, table_name, column_schema.data_type, data_type,
+                ),
+            }
+        );
+
+        Ok(())
+    }
+
+    fn try_get_data_type_from_value(value: &PbValue) -> Result<DatumKind> {
+        match value {
+            PbValue::Float64Value(_) => Ok(DatumKind::Double),
+            PbValue::StringValue(_) => Ok(DatumKind::String),
+            PbValue::Int64Value(_) => Ok(DatumKind::Int64),
+            PbValue::Float32Value(_) => Ok(DatumKind::Float),
+            PbValue::Int32Value(_) => Ok(DatumKind::Int32),
+            PbValue::Int16Value(_) => Ok(DatumKind::Int16),
+            PbValue::Int8Value(_) => Ok(DatumKind::Int8),
+            PbValue::BoolValue(_) => Ok(DatumKind::Boolean),
+            PbValue::Uint64Value(_) => Ok(DatumKind::UInt64),
+            PbValue::Uint32Value(_) => Ok(DatumKind::UInt32),
+            PbValue::Uint16Value(_) => Ok(DatumKind::UInt16),
+            PbValue::Uint8Value(_) => Ok(DatumKind::UInt8),
+            PbValue::TimestampValue(_) => Ok(DatumKind::Timestamp),
+            PbValue::VarbinaryValue(_) => Ok(DatumKind::Varbinary),
+        }
     }
 }
